Pull request #13 notes fixed
