+++ app/controllers/users_controller.rb
@@ -2,8 +2,11 @@
 
 # Users Controller class
 class UsersController < ApplicationController
-  require './app/lib/helpers'
   include Helpers
+
+  get '/login' do
+    erb :login
+  end
 
   get '/users/new' do
     @user = User.new({ id: 0, username: '', email: '', password_digest: '', active: false, is_admin: false })
